+++ src/settings/FileExtensions/FileExtensions.js
@@ -75,6 +75,8 @@ export class FileExtensions extends Component {
     mutator: PropTypes.shape({
       records: PropTypes.shape({
         POST: PropTypes.func.isRequired,
+        PUT: PropTypes.func.isRequired,
+        DELETE: PropTypes.func.isRequired,
       }).isRequired,
       restoreDefaultFileExtensions: PropTypes.shape({
         POST: PropTypes.func.isRequired,
@@ -128,13 +130,81 @@ export class FileExtensions extends Component {
 
     return records.POST(record)
       .catch(error => {
-        this.showCreateRecordErrorMessage(error, record);
+        this.showUpdateRecordErrorMessage(error, record);
 
         throw error;
       });
   };
 
-  handleCreateRecordSuccess = (record, dispatch, props) => {
+  editRecord = record => {
+    const { mutator: { records } } = this.props;
+
+    return records.PUT(record)
+      .catch(error => {
+        this.showUpdateRecordErrorMessage(error, record, true);
+
+        throw error;
+      });
+  };
+
+  deleteRecord = async record => {
+    const { mutator: { records } } = this.props;
+
+    try {
+      await records.DELETE(record);
+
+      const { match: { path } } = this.props;
+
+      this.transitionToParams({
+        _path: `${path}/view`,
+        layer: null,
+      });
+      this.showDeleteRecordSuccessfulMessage(record);
+    } catch (error) {
+      this.showDeleteRecordErrorMessage(record);
+    }
+  };
+
+  showDeleteRecordSuccessfulMessage = record => {
+    const message = (
+      <FormattedMessage
+        id="ui-data-import.settings.fileExtension.action.success"
+        values={{
+          extension: record.extension,
+          action: (
+            <strong>
+              <FormattedMessage id="ui-data-import.deleted" />
+            </strong>
+          ),
+        }}
+      />
+    );
+
+    this.callout.sendCallout({ message });
+  };
+
+  showDeleteRecordErrorMessage = record => {
+    const message = (
+      <FormattedMessage
+        id="ui-data-import.settings.fileExtension.action.error"
+        values={{
+          extension: record.extension,
+          action: (
+            <strong>
+              <FormattedMessage id="ui-data-import.deleted" />
+            </strong>
+          ),
+        }}
+      />
+    );
+
+    this.callout.sendCallout({
+      type: 'error',
+      message,
+    });
+  };
+
+  handleUpdateRecordSuccess = (record, dispatch, props) => {
     const { reset: resetForm } = props;
 
     resetForm();
@@ -147,13 +217,13 @@ export class FileExtensions extends Component {
     });
   };
 
-  async showCreateRecordErrorMessage(response, fileExtension) {
+  async showUpdateRecordErrorMessage(response, fileExtension, isEditMode = false) {
     const { extension } = fileExtension;
     const errorMsgIdEnding = await getXHRErrorMessage(response);
 
     const errorMsgId = errorMsgIdEnding
       ? `ui-data-import.validation.${errorMsgIdEnding}`
-      : 'ui-data-import.settings.fileExtension.create.error.network';
+      : `ui-data-import.settings.fileExtension.${isEditMode ? 'edit' : 'create'}.error.network`;
 
     const errorMessage = (
       <FormattedMessage
@@ -170,7 +240,7 @@ export class FileExtensions extends Component {
 
   renderActionMenu = menu => (
     <Button
-      data-test-restore-default-file-extensions
+      data-test-restore-default-file-extensions-button
       buttonStyle="dropdownItem"
       onClick={() => this.handleRestoreDefaultFileExtensions(menu)}
     >
@@ -226,7 +296,7 @@ export class FileExtensions extends Component {
     }
   };
 
-  createNewRecordContainerRef = ref => { this.newRecordContainer = ref; };
+  createFullWidthContainerRef = ref => { this.fullWidthContainer = ref; };
 
   createCalloutRef = ref => { this.callout = ref; };
 
@@ -275,18 +345,21 @@ export class FileExtensions extends Component {
                   updatedBy: intl.formatMessage({ id: 'ui-data-import.updatedBy' }),
                 }}
                 columnWidths={this.columnWidths}
-                newRecordContainer={this.newRecordContainer}
+                fullWidthContainer={this.fullWidthContainer}
                 ViewRecordComponent={ViewFileExtension}
                 EditRecordComponent={FileExtensionForm}
                 newRecordInitialValues={newRecordInitialValues}
                 showSingleResult={showSingleResult}
                 onCreate={this.createRecord}
-                handleCreateSuccess={this.handleCreateRecordSuccess}
+                onEdit={this.editRecord}
+                onDelete={this.deleteRecord}
+                handleCreateSuccess={this.handleUpdateRecordSuccess}
+                handleEditSuccess={this.handleUpdateRecordSuccess}
               />
             </div>
             <div
-              className={css.newRecordContainer}
-              ref={this.createNewRecordContainerRef}
+              className={css.fullWidthContainer}
+              ref={this.createFullWidthContainerRef}
             />
             <ConfirmationModal
               id="restore-default-file-extensions-modal"